requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb==0.5.3
python-dotenv